Plumb rest config through from manager. (#1334)
Plumb the config through alongside the client object (since they convey
the same information really) all the way to pipe task.

This way we use the same explicit config everywhere, rather than rely on
config.GetConfig to guess the correct one. This is especially important
when running the controller outside of a cluster.

Signed-off-by: Thomas Runyon <runyontr@gmail.com>